Much better handling of root commands (temporary shell script not needed anymore). Added root permission check before executing commands - this will give a much better message to users when root access cannot be acquired.
git-svn-id: https://droidwall.googlecode.com/svn/trunk@19 f190bac4-7986-11de-adf2-cbd3851e0ccf

+++ src/com/googlecode/droidwall/MainActivity.java
@@ -131,7 +131,8 @@ public class MainActivity extends ListActivity implements OnCheckedChangeListene
         	handler = new Handler() {
         		public void handleMessage(Message msg) {
         			if (progress != null) progress.dismiss();
-        			Api.showIptablesRules(MainActivity.this);
+        			if (!Api.hasRootAccess(MainActivity.this)) return;
+           			Api.showIptablesRules(MainActivity.this);
         		}
         	};
 			handler.sendEmptyMessageDelayed(0, 200);
@@ -140,10 +141,11 @@ public class MainActivity extends ListActivity implements OnCheckedChangeListene
     		progress = ProgressDialog.show(this, "Working...", "Applying iptables rules.", true);
         	handler = new Handler() {
         		public void handleMessage(Message msg) {
+        			if (progress != null) progress.dismiss();
+        			if (!Api.hasRootAccess(MainActivity.this)) return;
         			if (Api.refreshIptables(MainActivity.this, true)) {
         				Toast.makeText(MainActivity.this, "Rules applied with success", Toast.LENGTH_SHORT).show();
         			}
-        			if (progress != null) progress.dismiss();
         		}
         	};
 			handler.sendEmptyMessageDelayed(0, 200);
@@ -152,10 +154,11 @@ public class MainActivity extends ListActivity implements OnCheckedChangeListene
     		progress = ProgressDialog.show(this, "Working...", "Deleting iptables rules.", true);
         	handler = new Handler() {
         		public void handleMessage(Message msg) {
+        			if (progress != null) progress.dismiss();
+        			if (!Api.hasRootAccess(MainActivity.this)) return;
         			if (Api.purgeIptables(MainActivity.this)) {
         				Toast.makeText(MainActivity.this, "Rules purged with success", Toast.LENGTH_SHORT).show();
         			}
-        			if (progress != null) progress.dismiss();
         		}
         	};
 			handler.sendEmptyMessageDelayed(0, 200);
